Guard against accidential pushes on base branches

+++ git.sh
@@ -9,6 +9,13 @@ function pull()
 
 function push()
 {
+    GIT_BRANCH="$(git branch --show-current)"
+    if [[ "$(echo "$GIT_BRANCH" | grep -cP '^(develop|main|master)$')" == '1' && "$1" != "$GIT_BRANCH" ]]; then
+        printf "\e[31mERROR: Yer on $GIT_BRANCH"'!\e[0m\n'
+        echo "If you really want to push to $GIT_BRANCH, show you mean it:"
+        echo "$ push $GIT_BRANCH"
+        return 1
+    fi
     git push "$([[ "$(echo $1 | grep -cP '^-*f(orce)?$')" != '0' ]] && echo '--force ')"origin "$(git branch --show-current)"
 }
